Reduce copy overhead when writing metric events
Previously, otf2::event::metric stored its metric values as pairs of
referenced metric members and metric values.  This incurred overhead
when writing the event to a trace because of the following reasons:

    1.  OTF2_EvtWriter_Metric expects two continuous arrays, one for
        metric values and one for the corresponding OTF2_TypeIDs.
        Each of them had to be created and copied from the pairs
        contained in a metric event.

    2.  Each type ID had to be looked up in the referenced metric
        member.

With this commit, otf2::event::metric stores both metric values and type
IDs directly in two continuous buffers which.  The contents of these
buffers is copied exactly once from the OTF2 interface; when writing
an event to a trace, no additional copy operations are required.

It is still possible to obtain a otf2::event::metric::value_container
from a metric; they are created on-demand an will still contain a
refercence to metric member.

NOTE: this commit makes *BREAKING CHANGES* as the constructor for metric
events has changed.