Allow construction of metric events without values
This makes it possible to  pre-initializate a metric event according to
its metric class.

Memory for metric values is allocated for each metric member referenced
by the metric class.  The values are left uninitialized, but their type
is set from the corresponding metric member.

The intended use case of this functionality is performance critical code
that needs to write the same event multiple times with different values.
Before, each time an event were to be written with updated values, a new
event had to be constructed as its metric  values were immutable.   With
this, one otf2::event::metric may be pre-initialized and have its values
updated on-demand, which can then be written directly.