Reduce copy overhead when writing metric events (#9)
Previously, otf2::event::metric stored its metric values as pairs of
referenced metric members and metric values.  This incurred overhead
when writing the event to a trace because of the following reasons:

    1.  OTF2_EvtWriter_Metric expects two continuous arrays, one for
        metric values and one for the corresponding OTF2_TypeIDs.
        Each of them had to be created and copied from the pairs
        contained in a metric event.

    2.  Each type ID had to be looked up in the referenced metric
        member.

With this commit, otf2::event::metric stores both metric values and type
IDs directly in two continuous buffers. The contents of these
buffers is copied exactly once from the OTF2 interface; when writing
an event to a trace, no additional copy operations are required within otf2xx.

It is still possible to obtain an otf2::event::metric::value_container
from a metric; they are created on-demand an will still contain a
refercence to metric member.

NOTE: this commit makes *BREAKING CHANGES* as the constructor for metric
events has changed.

Now metric events can be pre-initializated according to its metric class.
Memory for metric values is allocated for each metric member referenced
by the metric class.  The values are left uninitialized, but their type
is set from the corresponding metric member.

The intended use case of this functionality is performance critical code
that needs to write the same event multiple times with different values.
Before, each time an event were to be written with updated values, a new
event had to be constructed as its metric  values were immutable.   With
this, one otf2::event::metric may be pre-initialized and have its values
updated on-demand, which can then be written directly.